refactoring: doing relational style columns
Damm, why are am I using non relational database. Here's the deal:

According to relational db standards, 1:m relations should be recorded
only on the 'many' side. The 'one' side dont knows what are its
relatives.
But Mongoose docs recomment to setup an array on the 'one' side too,
which puts a lot of effort on keeping this info tracked in every single
delete and create. So I'm ignoring Mongoose docs and doing it in the
relational style.

Probably I'm wrong, though